+++ R/mrs_data_proc.R
@@ -1362,6 +1362,37 @@ seconds <- function(mrs_data) {
   seq(from = 0, to = (Npts(mrs_data) - 1) / fs, by = 1 / fs)
 }
 
+#' Return a time scale vector of acquisition times for a dynamic MRS scan. The
+#' first temporal scan is assigned a value of 0.
+#' @param mrs_data MRS data.
+#' @return time scale vector in units of seconds.
+#' @export
+dyn_acq_times <- function(mrs_data) {
+  
+  if (is.na(tr(mrs_data)) | is.null(tr(mrs_data))) {
+    stop("TR not set, use set_tr function to set the repetition time.")
+  }
+  
+  if (is.na(Ntrans(mrs_data)) | is.null(Ntrans(mrs_data))) {
+    stop("Number of transients not set, use set_Ntrans function to set the 
+         number of transients.")
+  }
+  
+  n_trans <- Ntrans(mrs_data)
+  TR      <- tr(mrs_data)
+  n_dyns  <- Ndyns(mrs_data)
+  t_acq   <- seq(from = 0, by = TR, length.out = n_trans)
+  
+  # correct for missmatch between n_trans and n_dyns due to temporal averaging 
+  if (n_trans != n_dyns) {
+    if (n_trans%%n_dyns != 0) stop("Dynamics and transients do not match")
+    block_size <- n_trans / n_dyns
+    t_acq <- colMeans(matrix(t_acq, nrow = block_size))
+  }
+  
+  return(t_acq)
+}
+
 #' Get the indices of data points lying between two values (end > x > start).
 #' @param scale full list of values.
 #' @param start smallest value in the subset.